pyproject.toml
[project]
name = "attendance-monitoring-system"
version = "0.1.0"
description = "Advanced facial feature extraction system for attendance monitoring"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core ML and Computer Vision
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    
    # Face Recognition and Detection
    "face-recognition>=1.3.0",
    "facenet-pytorch>=2.5.3",
    "insightface>=0.7.3",
    "retinaface-pytorch>=0.0.3",
    
    # Advanced Feature Extraction
    "mediapipe>=0.10.0",
    "dlib>=19.24.0",
    "mtcnn>=0.1.1",
    
    # Deep Learning Frameworks
    "transformers>=4.30.0",
    "timm>=0.9.0",
    
    # Image Processing and Augmentation
    "albumentations>=1.3.0",
    "scikit-image>=0.21.0",
    
    # Data Management
    "pandas>=2.0.0",
    "pymongo>=4.0.0",
    "motor>=3.0.0",
    
    # API and Web Framework
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    
    # Utilities
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
